requirements.txt
SpeechRecognition==3.8.1
Pillow==8.2.0
nltk==3.6.2
requests==2.25.1
beautifulsoup4==4.9.3